# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Boilerplate for a Go REST web service.",
        "title": "quick-rest",
        "contact": {
            "name": "Huy Nguyen"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/items": {
            "get": {
                "description": "Get all items.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Get all items.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Item"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create an item.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Create an item.",
                "parameters": [
                    {
                        "description": "Create an item",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateItem"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/items/{id}": {
            "get": {
                "description": "Get an item by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Get an item by ID.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Item"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an item.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Delete an item.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "patch": {
                "description": "Update an item.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "items"
                ],
                "summary": "Update an item.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update an item",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CreateItem": {
            "type": "object",
            "required": [
                "attr1",
                "attr2"
            ],
            "properties": {
                "attr1": {
                    "type": "string"
                },
                "attr2": {
                    "type": "number"
                }
            }
        },
        "model.Item": {
            "type": "object",
            "properties": {
                "attr1": {
                    "type": "string"
                },
                "attr2": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "model.UpdateItem": {
            "type": "object",
            "properties": {
                "attr1": {
                    "type": "string"
                },
                "attr2": {
                    "type": "number"
                }
            }
        }
    }
}